Editing function arguments
Switching the arguments from a 'double' to an 'int'. This will resolve the "Infinity" issue with the factorial logic.

+++ GitDemo.java
@@ -25,7 +25,7 @@ public class GitDemo {
      * i.e. if num = 4,
      *  the function will return the value of (1 * 2 * 3 * 4)
      *****/
-    public double findFactorial(double num) {
+    public double findFactorial(int num) {
         double newNumber = 1;
         
         for(int i = 1; i <= num; ++i){
@@ -40,7 +40,7 @@ public class GitDemo {
      * i.e. if num = 1234
      *  the function will return 4321
      *****/
-    public double reverseNumber(double num){
+    public double reverseNumber(int num){
         double factorialNumber = 0;
         
         while( num != 0 ) {
